Refactored OAuth sign in buttons to own components
